[config] Created a ConfigurationManager
This change affects the whole project since it now allows the targets to
interact directly with the configuration file. Basicaly, a Target now
inherits from a ConfigurationManager and all the class methods
interacting with the configuration file now use the methods of the
superclass.

This modifications make sense since we only allow the targets (and maybe in the
future, filters) to interact with the configuration file.
Also, there were a lot of useless methods in this file.
