+++ hapycolor/hapycolor.config
@@ -1,8 +1,4 @@
-[core]
-app_name = hapycolor
-
 [hyperplan]
-keys = ./hyperplan/keys.json
 bright = ./hyperplan/hyperplans_v2.json
 dark = ./hyperplan/hyperplans_v2.json
 saturation = ./hyperplan/saturations.json
@@ -12,10 +8,6 @@ iterm_template = ./hapycolor/targets/iterm_template.xml
 
 [Wallpaper]
 wallpaper_macos = ~/Library/Application Support/Dock/desktoppicture.db
-
-[Filters]
-lum_filter = 2
-reducer = 5
 
 [Gnome]
 gnome_terminal_profiles = /org/gnome/terminal/legacy/profiles:
