+++ hapycolor/targets/vim.py
@@ -1,4 +1,4 @@
-from hapycolor import config
+from hapycolor import targets
 from hapycolor import helpers
 from hapycolor import exceptions
 from hapycolor import palette as pltte
@@ -87,7 +87,7 @@ class Vim(base.Target):
         Vim.save_config({Vim.configuration_key: (p / "hapycolor.vim").as_posix()})
 
     def custom_path():
-        p = config.input_path("Path to vim's custom plugins directory: ")
+        p = helpers.input_path("Path to vim's custom plugins directory: ")
         if not p.is_absolute():
             p = p.resolve()
 
@@ -95,7 +95,7 @@ class Vim(base.Target):
             raise exceptions.WrongInputError("Must be a directory")
 
     def compatible_os():
-        return [config.OS.LINUX, config.OS.DARWIN]
+        return [targets.OS.LINUX, targets.OS.DARWIN]
 
     def export(palette, image_path):
         if palette.__class__ != pltte.Palette or not palette.is_initialized:
@@ -115,7 +115,7 @@ set background=dark
 if exists("syntax_on")
 syntax reset
 endif
-let g:colors_name = "%s"''' % config.app_name()
+let g:colors_name = "Hapycolor"'''
         f.write(header + "\n\n")
 
     @staticmethod
