+++ tests/test_targets.py
@@ -1,6 +1,7 @@
 from hapycolor import config
 from hapycolor import targets
-from hapycolor.targets import base, vim, lightline, iterm, wallpaper
+from hapycolor import exceptions
+from hapycolor.targets import base, vim, lightline, iterm, wallpaper, gnome, i3
 from tests import helpers
 from unittest.mock import patch
 import unittest
@@ -22,15 +23,15 @@ class TestTarget(unittest.TestCase):
             except Exception as e:
                 self.fail(str(e))
 
-    @patch('hapycolor.config.os', return_value=config.OS.LINUX)
+    @patch('hapycolor.targets.os', return_value=targets.OS.LINUX)
     def test_get_compatible_linux(self, mock_os):
         for t in targets.get_compatible():
-            self.assertIn(config.OS.LINUX, t.compatible_os())
+            self.assertIn(targets.OS.LINUX, t.compatible_os())
 
-    @patch('hapycolor.config.os', return_value=config.OS.DARWIN)
+    @patch('hapycolor.targets.os', return_value=targets.OS.DARWIN)
     def test_get_compatible_darwin(self, mock_os):
         for t in targets.get_compatible():
-            self.assertIn(config.OS.DARWIN, t.compatible_os())
+            self.assertIn(targets.OS.DARWIN, t.compatible_os())
 
     def test_reconfigure_valid_target(self):
         try:
@@ -56,8 +57,9 @@ class TestTarget(unittest.TestCase):
             `targets.__all__`, defined in this test.
         """
         import inspect
-        targets.__all__ = ['vim', 'lightline', 'iterm', 'wallpaper']
-        target_modules = inspect.getmembers(targets)[0][1]
+        targets.__all__ = ['vim', 'lightline', 'iterm', 'wallpaper', 'gnome',
+                           'i3']
+        target_modules = inspect.getmembers(targets)[1][1]
         for m in target_modules:
             try:
                 self.assertTrue(inspect.ismodule(eval("targets." + m)))
@@ -67,3 +69,26 @@ class TestTarget(unittest.TestCase):
                 self.assertIsInstance(clazz(), base.Target)
             except NameError as e:
                 self.fails(str(e))
+
+    @patch('platform.system', return_value="Windows")
+    def test_not_supported_operating_system(self, get_os):
+        """ Checks that the 'os' getter works properly when running the program
+            on an unsupported platform """
+        with self.assertRaises(exceptions.PlatformNotSupportedError):
+            targets.os()
+
+    @patch('platform.system', side_effect=["Linux", "Darwin"])
+    def test_supported_operating_system(self, get_os):
+        """ Checks that the 'os' getter works properly when running the program
+            on a supported platform """
+        try:
+            self.assertEqual(targets.os(), targets.OS.LINUX)
+        except exceptions.PlatformNotSupportedError as err:
+            self.fail("'os' getter raised error: " + str(err) + " - on a Linux"
+                      + " platform")
+
+        try:
+            self.assertEqual(targets.os(), targets.OS.DARWIN)
+        except exceptions.PlatformNotSupportedError as err:
+            self.fail("'os' getter raised error: " + str(err) + " - on a"
+                      + " Darwin platform")
